Revert "修复bug,支持win10"
This reverts commit fcd65bed3b377c6fd549feac2f2b8aa67db65334.

+++ Stub/Stub.h
@@ -23,7 +23,7 @@ extern "C"
 		BYTE        byXor;                 // 加密异或的值
 	}GLOBAL_PARAM, *PGLOBAL_PARAM;
 
-	typedef ULONGLONG(WINAPI *fnGetProcAddress)(_In_ HMODULE hModule, _In_ LPCSTR lpProcName);
+	typedef DWORD(WINAPI *fnGetProcAddress)(_In_ HMODULE hModule, _In_ LPCSTR lpProcName);
 	typedef HMODULE(WINAPI *fnLoadLibraryA)(_In_ LPCSTR lpLibFileName);
 	typedef int(WINAPI *fnMessageBox)(HWND hWnd, LPSTR lpText, LPSTR lpCaption, UINT uType);
 	typedef HMODULE(WINAPI *fnGetModuleHandleA)(_In_opt_ LPCSTR lpModuleName);
